configuração de autenticação do front end com back e Upload

+++ capsula-do-tempo/server/src/routes/memories.ts
@@ -1,39 +1,51 @@
 import { FastifyInstance } from 'fastify'
-import {z} from "zod"
+import { z } from 'zod'
 import { prisma } from '../lib/prisma'
 
 export async function memoriesRoutes(app: FastifyInstance) {
-  app.get('/memories', async () => {
-   const memories = await prisma.memory.findMany({
-    orderBy:{
-        createdAt: 'asc',
-    }
-   }),
+  app.addHook('preHandler', async (request) => {
+    await request.jwtVerify()
+  })
 
-   return memories.map(memory => {
-    return{
+  app.get('/memories', async (request) => {
+    await request.jwtVerify()
+
+    const memories = await prisma.memory.findMany({
+      where: {
+        userId: request.user.sub,
+      },
+      orderBy: {
+        createdAt: 'asc',
+      },
+    })
+
+    return memories.map((memory) => {
+      return {
         id: memory.id,
         coverUrl: memory.coverUrl,
-        excerpt: memory.content.substring(0, 115).concat(...)
+        excerpt: memory.content.substring(0, 115).concat('...'),
+      }
+    })
+  })
+
+  app.get('/memories/:id', async (request, reply) => {
+    const paramsSchema = z.object({
+      id: z.string().uuid(),
+    })
+
+    const { id } = paramsSchema.parse(request.params)
+
+    const memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    })
+
+    if (!memory.isPlublic && memory.userId !== request.user.sub) {
+      return reply.status(401).send()
     }
-   })
-  })
 
-  app.get('/memories/:id', async (request) => {
-   const paramsSchema = z.object({
-    id: z.string().uuid(),
-   })
-
-   const {id} = paramsSchema.parse(request.params)
-
-   const memory = await prisma.memory.findUniqueOrThrow({
-    where: {
-      id,
-    },
-  })
-
-  return memory
-
+    return memory
   })
 
   app.post('/memories', async (request) => {
@@ -41,66 +53,81 @@ export async function memoriesRoutes(app: FastifyInstance) {
       content: z.string().uuid(),
       coverUrl: z.string(),
       isPlublic: z.coerce.boolean().default(false),
-     })
+    })
 
-    const { content, coverUrl, isPlublic} = bodySchema.parse(request.body)
+    const { content, coverUrl, isPlublic } = bodySchema.parse(request.body)
 
     const memory = await prisma.memory.create({
       data: {
         content,
         coverUrl,
         isPlublic,
-        userId: '0d9e1ddb-fcd4-47ee-b726-7db179e3cd2b'
+        userId: request.user.sub,
       },
     })
-    
+
     return memory
   })
 
-  app.put('/memories/:id', async (request) => {
+  app.put('/memories/:id', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
-     })
-     const {id} = paramsSchema.parse(request.params)
+    })
+    const { id } = paramsSchema.parse(request.params)
 
     const bodySchema = z.object({
       content: z.string().uuid(),
       coverUrl: z.string(),
       isPlublic: z.coerce.boolean().default(false),
-     })
+    })
 
-    const { content, coverUrl, isPlublic} = bodySchema.parse(request.body)
+    const { content, coverUrl, isPlublic } = bodySchema.parse(request.body)
 
-    const memory = await.prisma.memory.update({
-      where:{
+    let memory = await prisma.memory.findUniqueOrThrow({
+      where: {
         id,
       },
-      data:{
-        content, 
-        coverUrl, 
+    })
+
+    if (memory.userId !== request.user.sub) {
+      return reply.status(401).send()
+    }
+
+    memory = await prisma.memory.update({
+      where: {
+        id,
+      },
+      data: {
+        content,
+        coverUrl,
         isPlublic,
-      }
+      },
     })
 
     return memory
-    
-   
   })
 
-  app.delete('/memories/:id', async (request) => {
+  app.delete('/memories/:id', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
-     })
-  
-     const {id} = paramsSchema.parse(request.params)
-  
+    })
+
+    const { id } = paramsSchema.parse(request.params)
+
+    const memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    })
+
+    if (memory.userId !== request.user.sub) {
+      return reply.status(401).send()
+    }
+
     await prisma.memory.delete({
       where: {
         id,
       },
     })
-  
-
-   
   })
 }
